w1/project-alpha/backend/pyproject.toml
[project]
name = "ticket-tag-management-backend"
version = "1.0.0"
description = "Backend API for Ticket Tag Management Tool"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "sqlalchemy==2.0.25",
    "psycopg2-binary==2.9.9",
    "alembic==1.13.1",
    "pydantic==2.5.3",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    "black==24.1.1",
    "isort==5.13.2",
    "flake8==7.0.0",
    "flake8-docstrings",
    "flake8-bugbear",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["app"]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503", "B008", "D100", "D103", "D104", "D106", "D107", "D400"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    ".mypy_cache",
    ".pytest_cache",
    "build",
    "dist",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
